docker
There were problems with the user.
I removed the non-root user in the docker
and added the user to the run command so that the
repository user conforms with the docker user.

Issue was that adding an application user generated a new user that was not compatible with the repository user.

Secondly the PDF cannot be shown from within the Docker.
I test if the program is run from Docker and if so tell
the user to look at the pdf locally.